# src/TorchWeb/OpenAPIs/openapi.json
{
  "components": {
    "schemas": {
      "CollectionResponse": {
        "properties": {
          "cardimg": {
            "items": {
              "$ref": "#/components/schemas/SpecimenImageResponse"
            },
            "type": "array"
          },
          "code": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "specimencount": {
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TorchTask"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SpecimenImageResponse": {
        "properties": {
          "external_url": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "SpecimenResponse": {
        "properties": {
          "card_image": {
            "$ref": "#/components/schemas/SpecimenImageResponse"
          },
          "catalog_number": {
            "type": "string"
          },
          "collection_code": {
            "type": "string"
          },
          "collection_id": {
            "type": "integer"
          },
          "collection_name": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/SpecimenImageResponse"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "SpecimensResponse": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "specimens": {
            "items": {
              "$ref": "#/components/schemas/SpecimenResponse"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TorchTask": {
        "properties": {
          "description": {
            "nullable": true,
            "type": "string"
          },
          "func_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "error_message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "TorchHub API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/collections/{collection_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Collection Delete",
        "tags": [
          "Collections"
        ]
      }
    },
    "/collections/{collectionid}": {
      "get": {
        "operationId": "GetCollection",
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Collection Get",
        "tags": [
          "Collections"
        ]
      }
    },
    "/collections/{collectionid}/csv": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Collection Export Csv",
        "tags": [
          "Collections"
        ]
      }
    },
    "/collections/{collectionid}/specimens": {
      "get": {
        "operationId": "GetSpecimens",
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search_string",
            "required": false,
            "schema": {
              "default": null,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecimensResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Collection Specimens",
        "tags": [
          "Collections"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Upload",
        "tags": [
          "Collections"
        ]
      }
    },
    "/collections/{collectionid}/specimens/{specimenid}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "specimenid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Specimen Delete",
        "tags": [
          "Collections"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "specimenid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Specimen Get",
        "tags": [
          "Collections"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "collectionid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "specimenid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Retry",
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
...
}